refactor: move dashboard into feature-based structure
