+++ frontend/src/features/dashboard/components/FraudGauge.tsx
