+++ frontend/src/features/dashboard/components/FraudReviewTable.tsx
@@ -1,4 +1,4 @@
-import DataTable from "./DataTable";
+import DataTable from "../../components/UI/DataTable";
 
 function FraudReviewTable() {
   // Dummy data — backend will replace this later
